Update my code 4

+++ basic_math.py
@@ -40,8 +40,9 @@ def get_median(number_list):
 
     number_list.sort()
     if length % 2:
-        median = number_list[length / 2]
+        median = number_list[int(length / 2)]
     else:
-        median = int(number_list[length / 2 - 1] + number_list[length / 2] / 2)
+        median = float(number_list[int(length / 2 - 1)] +
+                       number_list[int(length / 2)]) / 2
 
     return median
